feat: sign up, and fix old pages
