+++ mobile/lib/widgets/custom_button.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 
+// ignore: must_be_immutable
 class CustomButton extends StatelessWidget {
   final Widget icon;
   final String label;
